refactor: move enum logic to EnumCollector
